Trait pallet_fruniques::pallet::Config

pub trait Config: Config + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RemoveOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ChildMaxLen: Get<u32>;
    type MaxParentsInCollection: Get<u32>;
    type Rbac: RoleBasedAccessControl<Self::AccountId>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type RemoveOrigin: EnsureOrigin<Self::RuntimeOrigin>

type ChildMaxLen: Get<u32>

Maximum number of children a Frunique can have

type MaxParentsInCollection: Get<u32>

Maximum number of roots a Collection can have

type Rbac: RoleBasedAccessControl<Self::AccountId>

The fruniques pallet id, used for deriving its sovereign account ID.